requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
jiter==0.8.0
multidict==6.1.0
openai==1.55.3
pathlib==1.0.1
propcache==0.2.0
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
twilio==9.3.7
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.0
